serialize show action for list and it's items.
Structure JSON so when lists are requested only all the lists for a particular
user appear. When a specific list is requested the list and its items are returned.
The JSON is formatted for
requested: lists return: list_id, list_name
requested: list return: list_id, list_name with nest of item_id, item